+++ packages/api-reference/src/features/Operation/layouts/ClassicLayout.vue
@@ -10,11 +10,6 @@ import {
   ScalarIconPlay,
   ScalarIconWebhooksLogo,
 } from '@scalar/icons'
-import {
-  getOperationStability,
-  getOperationStabilityColor,
-  isOperationDeprecated,
-} from '@scalar/oas-utils/helpers'
 import { useClipboard } from '@scalar/use-hooks/useClipboard'
 import { getResolvedRef } from '@scalar/workspace-store/helpers/get-resolved-ref'
 import type { SecuritySchemeObjectSecret } from '@scalar/workspace-store/request-example'
@@ -35,6 +30,11 @@ import { ExternalDocs } from '@/features/external-docs'
 import Callbacks from '@/features/Operation/components/callbacks/Callbacks.vue'
 import OperationParameters from '@/features/Operation/components/OperationParameters.vue'
 import OperationResponses from '@/features/Operation/components/OperationResponses.vue'
+import {
+  getOperationStability,
+  getOperationStabilityColor,
+  isOperationDeprecated,
+} from '@/features/Operation/helpers/operation-stability'
 import type { OperationProps } from '@/features/Operation/Operation.vue'
 import {
   REQUEST_BODY_COMPOSITION_INDEX_SYMBOL,
@@ -238,8 +238,8 @@ const { copyToClipboard } = useClipboard()
         <!-- Request Example -->
         <ScalarErrorBoundary>
           <OperationCodeSample
-            v-model:selectedExample="selectedExampleKey"
             :key="requestBodyCompositionSelectionKey"
+            v-model:selectedExample="selectedExampleKey"
             class="operation-example-card"
             :clientOptions
             :eventBus
