+++ packages/api-reference/src/features/example-responses/helpers/has-response-content.test.ts
