+++ packages/api-reference/src/features/example-responses/helpers/normalize-mime-type-object.test.ts
